Linuxビルドの修正。

+++ NVEnc/frm/frmConfig.cpp
@@ -633,7 +633,7 @@ ToolStripMenuItem^ frmConfig::fcgTSSettingsSearchItem(String^ stgPath) {
 System::Void frmConfig::SaveToStgFile(String^ stgName) {
     size_t nameLen = CountStringBytes(stgName) + 1;
     TCHAR *stg_name = (TCHAR *)malloc(nameLen * sizeof(TCHAR));
-    GetWCHARfromString(stg_name, nameLen, stgName);
+    GetWCHARfromString(stg_name, (DWORD)nameLen, stgName);
     init_CONF_GUIEX(cnf_stgSelected, FALSE);
     FrmToConf(cnf_stgSelected);
     String^ stgDir = Path::GetDirectoryName(stgName);
